pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shoshin"
dynamic = ["version"]
description = 'Conversational AI that gets context from a video course and provides teaching support to students, with an attitude of openness, eagerness, and lack of preconceptions, just as a beginner would.'
readme = "README.md"
requires-python = ">=3.7"
license = "Apache-2.0"
keywords = []
authors = [
  { name = "Emanuele Palazzetti", email = "emanuele.palazzetti@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "ffmpeg-python",
]

[project.optional-dependencies]
dev = [
  "mypy",
  "pre-commit",
  # Test
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "tox",
]

all = [
  "shoshin[dev]",
]

[project.scripts]
shoshin = "shoshin.cli.commands:cli"

[project.urls]
Documentation = "https://github.com/palazzem/shoshin#readme"
Issues = "https://github.com/palazzem/shoshin/issues"
Source = "https://github.com/palazzem/shoshin"

[tool.hatch.version]
path = "src/shoshin/__about__.py"
